Create separate components for the two input fields
